Engram-AiMemory/01_devroot/RemoteAgent/dev/python/requirements.txt
# Core
weaviate-client>=4.4.0
watchdog>=4.0.0
python-dotenv>=1.0.0

# LM Studio SDK
lmstudio-sdk>=0.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
striprtf>=0.0.26
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Chunking & Text Processing
tiktoken>=0.5.0
langchain-text-splitters>=0.0.1

# API Server (for TS orchestration)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Utilities
rich>=13.7.0
httpx>=0.26.0
aiofiles>=23.2.0
xxhash>=3.4.0
